+++ tests/test_hu.py
@@ -1,9 +1,9 @@
-from hw import __version__, ObjectDict
+from hu import __version__, ObjectDict
 import toml
 
 
 def test_version():
-    "Ensure that hw reports the version set in pyproject.toml."
+    "Ensure that hu reports the version set in pyproject.toml."
     # Should this test only be a pre-merge hook or similar?
     with open("pyproject.toml") as f:
         # ``we now eat our own dogfood for the first time
@@ -13,6 +13,6 @@ def test_version():
 
 def test_old_import():
     "Verify that a backwards-compatible import still works."
-    from hw.object_dict import ObjectDict as OD
+    from hu.object_dict import ObjectDict as OD
 
     assert OD is ObjectDict
