+++ tox.ini
@@ -9,5 +9,5 @@ base_python =
 [testenv]
 whitelist_externals = poetry
 commands =
-    /Users/sholden/bin/poetry install -v
-    /Users/sholden/bin/poetry run pytest tests/
+    poetry install -v
+    pytest -v --cov=src  --cov-report=term-missing
